Deprecate local() from font face src property
Since there's no way to guarantee that a local font file named helvetica is actually helvetica or the helvetica that we want, I've decided to deprecate it.

Most of the reasons are well explained here, more than i can do in a commit message: https://www.bramstein.com/writing/web-font-anti-patterns-local-fonts.html

Nonetheless, added a link to a gist where you can see a how to achieve the most optimal way of using local() based on the font name.